Analysis (#19)
* Revert "2019/1/12 스터디 노트"

This reverts commit 8c28be6f5864e4d748113f02e9956d70ab1a4522.

* Revert "2019/1/12 스터디노트"

This reverts commit 11fbf5f4f36320ebcf7b1d17bb967237789cb7e6.

* 19.01.12 (커널공부, JL, 오곡라뗴, 아자아자파이팅, 바뿌사, 코딩의노예)

* 1/19 스터디 (코딩코딩, 코딩의 노예, 아자아자파이팅, 커널공부, 휴프)

+++ SET_TAGS_CSCOPE.sh
@@ -1,5 +1,9 @@
-rm tags
-rm cscope.*
+if [ -f "tags" ]; then
+	rm tags
+fi
+if [ -f "cscope.*" ]; then
+	rm cscope.*
+fi
 
 ctags -aR --extra=f *
